requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
torch
imbalanced-learn
uvicorn[standard]
fastapi
SQLPyHelper
pydantic
joblib
python-dotenv
streamlit
requests